Update civic.json file to new specification
Hi, there! Code for DC is moving to an updated civic.json
specification. This pull request formats your existing
civic.json to the new standard.

Feel free to look it over and make any updates.

========================

There are a few errors remaining in the civic.json,
which will need a correction:

Error in contact: '' is too short

========================

The new civic.json spec keeps the required fields from the older
one, but has a number of benefits. It:

* removes fields that are hard to maintain and get out of date quickly
* combines partner fields to be more broadly applicable
* is usable by civic tech projects outside of government, as well as inside

You'll see that DC Government is using this standard in its own repos:

https://github.com/dcgov

You can learn more about the specification requirements here:

http://open.dc.gov/civic.json